requirements.txt
PyQt5>=5.15
python-vlc>=3.0
requests>=2.0

# Testing dependencies
pytest>=7.0
pytest-qt>=4.0
pytest-mock>=3.0
pytest-cov>=4.0